move SymGrad and RCall kernels to functional_ops
This moving of SymbolicGradient and RemoteCall is because
  their OPs are defined in ops/functional_ops.h .

Benefits:
if a C++ developer only imports ops/function_ops.cc and kernels/function_ops.cc,
  without 2 functional_ops.cc files,
then there won't be runtime errors reporing the OPs are not found.

Other influence:
This commit removes the 2 kernels from //tensorflow/core/kernels:android_core_ops,
  which is safe since the two OPs are not defined yet.
